ShamIn/requirements.txt
# ===== ShamIn - SYP/USD Exchange Rate Forecaster =====
# Python 3.11+

# === Core ML (from ACTION-main) ===
torch>=2.3.0
pytorch-lightning>=2.2.0
pytorch-forecasting>=1.0.0
optuna>=3.5.0
pandas>=2.0
numpy>=1.26.0
scikit-learn>=1.4.0
xgboost>=2.0.3
statsmodels>=0.14.1

# === Data Collection ===
telethon>=1.34.0           # Telegram channels
feedparser>=6.0.10         # RSS feeds
beautifulsoup4>=4.12.0     # Web scraping
selenium>=4.16.0           # Dynamic scraping
lxml>=5.1.0                # HTML/XML parser
aiohttp>=3.9.0             # Async HTTP

# === Arabic NLP ===
arabic-reshaper>=3.0.0
python-bidi>=0.4.2
camel-tools>=1.5.2
farasasegmenter>=0.0.14

# === Infrastructure ===
psycopg2-binary>=2.9.9     # PostgreSQL
influxdb-client>=1.39.0    # InfluxDB
redis>=5.0.1               # Redis
minio>=7.2.3               # MinIO object storage
sqlalchemy>=2.0.25          # ORM
alembic>=1.13.0            # DB migrations

# === Task Queue ===
celery>=5.3.4
celery[redis]

# === API & Dashboard ===
fastapi>=0.109.0
uvicorn>=0.27.0
streamlit>=1.30.0
plotly>=5.18.0
pydantic>=2.5.0

# === Monitoring & MLOps ===
mlflow>=2.10.0
shap>=0.44.0

# === Utilities ===
PyYAML>=6.0.1
python-dotenv>=1.0.0
loguru>=0.7.2
tqdm>=4.66.0
requests>=2.31.0

# === Advanced (Phase 8+) ===
transformers>=4.37.0       # LLMs
